server: move inspectEngines() to init.go where it belongs
Release note: None
